requirements.txt
streamlit
pandas
PyMuPDF # For PyMuPDF (fitz)
Pillow  # For image/logo display
python-dateutil  #parser and relativedelta

